+++ endpoints/cronjobs/sendcancellationnotifications.php
@@ -320,13 +320,14 @@ while ($userToNotify = $usersToNotify->fetchArray(SQLITE3_ASSOC)) {
                         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
 
                         $response = curl_exec($ch);
-                        curl_close($ch);
-
+                        
                         if ($response === false) {
                             echo "Error sending notifications: " . curl_error($ch) . "<br />";
                         } else {
                             echo "Discord Notifications sent<br />";
                         }
+                        
+                        unset($ch);
                     }
                 }
             }
@@ -468,13 +469,13 @@ while ($userToNotify = $usersToNotify->fetchArray(SQLITE3_ASSOC)) {
 
                     $result = curl_exec($ch);
 
-                    curl_close($ch);
-
                     if ($result === false) {
                         echo "Error sending notifications: " . curl_error($ch) . "<br />";
                     } else {
                         echo "Pushover Notifications sent<br />";
                     }
+                    
+                    unset($ch);
                 }
             }
 
@@ -523,13 +524,14 @@ while ($userToNotify = $usersToNotify->fetchArray(SQLITE3_ASSOC)) {
                         }
 
                         $response = curl_exec($ch);
-                        curl_close($ch);
-
+                        
                         if ($response === false) {
                             echo "Error sending notifications: " . curl_error($ch) . "<br />";
                         } else {
                             echo "Ntfy Notifications sent<br />";
                         }
+                        
+                        unset($ch);
                     }
                 }
             }
@@ -586,13 +588,14 @@ while ($userToNotify = $usersToNotify->fetchArray(SQLITE3_ASSOC)) {
                             // Execute the cURL request
                             $response = curl_exec($ch);
                             $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-                            curl_close($ch);
                 
                             if ($response === false || $httpCode >= 400) {
                                 echo "Error sending cancellation notifications: " . curl_error($ch) . "<br />";
                             } else {
                                 echo "Webhook Cancellation Notification sent for subscription: " . $subscription['name'] . "<br />";
                             }
+                            
+                            unset($ch);
                 
                             usleep(1000000); // 1s delay between requests
                         }