+++ includes/checksession.php
@@ -1,6 +1,12 @@
 <?php
 // Handle OIDC first
+$secondsInMonth = 30 * 24 * 60 * 60;
 if (session_status() === PHP_SESSION_NONE) {
+    session_set_cookie_params([
+        'lifetime' => $secondsInMonth,             
+        'httponly' => true,          
+        'samesite' => 'Lax'          
+    ]);
     session_start();
 }